requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Werkzeug==2.3.7
PyMuPDF==1.23.8
python-docx==1.1.0
pytesseract==0.3.10
Pillow==10.2.0
python-dotenv==1.0.0
requests==2.31.0

# AI Features (Optional - install separately if needed)
# Uncomment and install these for AI-powered question analysis:
# sentence-transformers==2.2.2
# faiss-cpu==1.7.0
# openai==1.12.0
